% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualization_multiplier.R
\name{annualization_multiplier}
\alias{annualization_multiplier}
\title{Retrieve the annualization multiplier}
\usage{
annualization_multiplier(period)
}
\description{
Specify either a character annualization period or a numeric annualization multiplier of
your own.
}
\details{
The following character annualization specifications are supported.
\itemize{
\item "day"     = 252
\item "week"    = 52
\item "month"   = 12
\item "quarter" = 4
\item "year"    = 1
}
}
\examples{

annualization_multiplier("day")
annualization_multiplier("2 day")
annualization_multiplier(250)

}
