% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_return.R
\name{cumulative_return}
\alias{cumulative_return}
\alias{total_return}
\title{Cumulative returns}
\usage{
cumulative_return(.r, type = "arithmetic")

total_return(.r, type = "arithmetic")
}
\arguments{
\item{.r}{A vector of returns.}

\item{type}{Either \code{"arithmetic"} or \code{"log"} returns.}
}
\description{
Various functions for aggregating returns. \code{cumulative_return()} calculates
a vector of cumulative returns. \code{total_return()} returns the total return
over the period.
}
\section{Implementation}{


Arithmetic returns are compounded as:

\deqn{(1 + r_1)(1 + r_2) ... (1 + r_n) - 1}

Log returns are compounded as:

\deqn{r_1 + r_2 + ... + r_n}

\code{total_return()} is the last value of the cumulative return.
}

