% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_supported_datasets.R
\name{list_supported_datasets_other}
\alias{list_supported_datasets_other}
\title{List Supported Other Datasets}
\usage{
list_supported_datasets_other()
}
\value{
A tibble with columns: \code{type} (the type of dataset), \code{dataset_name}
(the name of the data source), and \code{domain} (the domain to which the
dataset belongs).
}
\description{
Returns a tibble listing the supported other datasets and their
corresponding dataset names.
}
\seealso{
Other utility functions: 
\code{\link{create_summary_statistics}()},
\code{\link{get_available_huggingface_files}()},
\code{\link{list_supported_datasets}()},
\code{\link{list_supported_datasets_ff}()},
\code{\link{list_supported_datasets_ff_legacy}()},
\code{\link{list_supported_datasets_macro_predictors}()},
\code{\link{list_supported_datasets_pseudo}()},
\code{\link{list_supported_datasets_wrds}()},
\code{\link{list_supported_indexes}()},
\code{\link{list_tidy_finance_chapters}()},
\code{\link{open_tidy_finance_website}()},
\code{\link{trim}()},
\code{\link{validate_dates}()},
\code{\link{winsorize}()}
}
\concept{utility functions}
