% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match-helpers.R
\name{matches_ohlca}
\alias{matches_ohlca}
\alias{matches_ohlcav}
\title{Helpers for matching common financial columns}
\usage{
matches_ohlca(vars = NULL)

matches_ohlcav(vars = NULL)
}
\description{
Extensions of the family of "select helpers" from \code{tidyselect} like
\code{\link[tidyselect:contains]{tidyselect::contains()}} and \code{\link[tidyselect:matches]{tidyselect::matches()}}. \code{matches_ohlca()} selects the
columns \code{open}, \code{high}, \code{low}, \code{close}, and \code{adjusted}. \code{matches_ohlcav()}
additionally matches \code{volume}.
}
\details{
As a user, you should rarely, if ever, need to specify \code{vars}.
}
\examples{

data(FANG)

dplyr::select(FANG, matches_ohlca())

FANG \%>\%
  as_tbl_time(date) \%>\%
  dplyr::group_by(symbol) \%>\%
  calculate_return(matches_ohlca())

}
