% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_return.R
\name{aggregate_return}
\alias{aggregate_return}
\alias{total_return}
\alias{cumulative_return}
\title{Aggregate returns}
\usage{
aggregate_return(.tbl_time, ..., type = "arithmetic", period = "daily",
  start_date = NULL)

total_return(.r, type = "arithmetic")

cumulative_return(.r, type = "arithmetic")
}
\arguments{
\item{.tbl_time}{A \code{tbl_time} object}

\item{...}{The columns to aggregate returns for.
One or more unquoted column names separated by commas.
Positive values select variables; negative values drop variables.}

\item{type}{Either \code{"arithmetic"} or \code{"log"} returns.}

\item{period}{A character specification used for time-based grouping. The
general format to use is \code{"frequency period"} where frequency is a number
like 1 or 2, and period is an interval like weekly or yearly. There must be
a space between the two.

Note that you can pass the specification in a flexible way:
\itemize{
\item 1 Year: \code{'1 year'} / \code{'1 Y'} / \code{'1 yearly'} / \code{'yearly'}
}

This shorthand is available for year, quarter, month, day, hour, minute,
and second periodicities. Note that "m" is the 1 letter specification used
for "month" and "M" is used for "minute". Otherwise case does not matter.}

\item{start_date}{Optional argument used to specify the start date for the
first group. The default is to start at the closest period boundary
below the minimum date in the supplied index.}

\item{.r}{A vector of returns.}
}
\description{
Various functions for aggregate returns. \code{total_return()} and \code{cumulative_return()}
aggregate a vector of returns with no regards for the period associated with
it. \code{aggregate_return()} aggregates returns in a \code{tbl_time} object with a
specified aggregation period (e.g., daily up to yearly return).
}
